Operators
Huge change. Operators allow for conditional or multi replaces and
required a rewrite of replace, evaluate and added params and status.
Status outputs a status, replace can handle multi replaces (almost)
evaluate is now purely a regex engine and params extracts operators and
such and feeds it to evaluate. Huge and awesome change